requirements.txt
# Runtime dependencies
pygame-ce>=2.4.0
Pillow>=9.0.0
requests>=2.28.0

# Packaging / web build helper
pygbag==0.9.2

# Test / CI tools
pytest>=7.0
playwright==1.55.0
pytest-timeout>=2.2.0
